+++ src/shared/IO/NativeAliases.h
@@ -6,17 +6,17 @@
 #include <Windows.h>
 #undef WIN32_LEAN_AND_MEAN
 
-namespace IO::Native {
-    typedef SOCKET SocketHandle;
+namespace IO { namespace Native {
+    typedef UINT_PTR SocketHandle;
     typedef HANDLE FileHandle;
-}
+}} // namespace IO::_Native
 
 #elif defined(__linux__)
 
-namespace IO::Native {
+namespace IO { namespace Native {
     typedef int SocketHandle;
     typedef int FileHandle;
-}
+}} // namespace IO::Native
 
 #endif
