+++ src/shared/IO/Networking/NetworkError.h
@@ -5,23 +5,29 @@
 
 namespace IO
 {
-    struct NetworkError {
+    class NetworkError {
+    public:
         enum class ErrorType {
             NoError,
             InternalError,
             SocketClosed,
             OnlyOneTransferPerDirectionAllowed,
         };
+    public:
+        explicit NetworkError(ErrorType errorType) : NetworkError(errorType, 0) {};
+        explicit NetworkError(ErrorType errorType, int osErrorCode) : m_error{errorType}, m_additionalOsErrorCode{osErrorCode} {};
 
-        ErrorType Error;
+        ErrorType GetErrorType() const { return m_error; };
+        int _GetOsErrorCode() const { return m_additionalOsErrorCode; };
 
-        explicit operator bool() const {
-            return Error != ErrorType::NoError;
-        };
+        explicit operator bool() const { return GetErrorType() != ErrorType::NoError; };
+        std::string ToString() const;
 
-        std::string toString() const {
-            return "TODO, Error to String";
-        }
+    private:
+        ErrorType m_error;
+        /// internal variable for toString(), might be os and situation dependent
+        int m_additionalOsErrorCode;
+
     };
 } // namespace IO
