+++ src/shared/IO/Networking/SocketDescriptor.h
@@ -1,10 +1,18 @@
 #ifndef MANGOS_IO_NETWORKING_SOCKETDESCRIPTOR_H
 #define MANGOS_IO_NETWORKING_SOCKETDESCRIPTOR_H
 
-#ifdef WIN32
-#include "./impl/windows/SocketDescriptor.h"
-#else
-#error "IO::Networking not supported on your platform"
-#endif
+#include <string>
+#include "./IpAddress.h"
+#include "../NativeAliases.h"
+
+namespace IO { namespace Networking {
+
+struct SocketDescriptor {
+public:
+    IO::Networking::IpEndpoint const m_peerEndpoint;
+    IO::Native::SocketHandle const _nativeSocket; // <-- Internal use only
+};
+
+}} // namespace IO::Networking
 
 #endif //MANGOS_IO_NETWORKING_SOCKETDESCRIPTOR_H
