+++ src/shared/IO/Networking/Utils.h
@@ -6,6 +6,7 @@
 
 namespace IO { namespace Networking
 {
+    /// Checks whenever the same (IPv4) Address is in the same subnet as the other one
     bool IsInSameSubnet(IpAddress const& ipAddressInQuestion, IpAddress const& subnetIpAddress, uint8_t subnetMaskInCidrNotation);
 }} // namespace IO::Networking
